provide injections at startup time in the safari extension
I've had a lot of trouble getting Safari extension localstorage data to be
accepted. If I write the file out when Safari isn't using the localstorage, the
values are dropped. If I write the file out while it is using the localstorage,
the database is locked. Sometimes I can write to the file, but I haven't been
able to consistently produce this. It doesn't help that any given MacOS update
might change how this works.

This commit changes the approach and instead generates a JavaScript file that is
loaded alongside the rest of the injector, and sets the injector's localstorage
at startup time. The approach isn't perfect and seems to require an arbitrary
"save" of the items, but once done for a single item, all items are applied.
This makes me think it needs a kick-off function called that applies the data
values that it has. I have documented these findings but have not executed on
them yet.