add fix for outdated homebrew casks
Older cask definitions use the "undent" method on String, which has been removed
from either later version of Homebrew or Ruby itself (I care not which applies).
This means upgrading from older definitions is problematic because the newer
executing (and self updating) Homebrew breaks against what was perfectly valid
definitions in a prior version. To get around this, the undent call is stripped
from where it is used and it is assumed that the string value in place will
suffice.

I have seen issues similar to this in the past, and I expect I will need to add
more repair scripting as I continue to use Homebrew.