fix spooky ghost snippets from loading
Somehow the really old version of the dndspell snippet was being loaded but I
couldn't find a reference to it anywhere on disk. The paths seemed correct, and
the .yas-compiled-snippets.el had the current version. I'm not sure where these
versions came from but they seem to be fixed now by forcing the custom yas
config to execute immediately. I suspect this is because the :config section
never actually gets run. Where the old snippet came from is still a mystery
though.