add means of disabling the sophos virus
The Sophos virus can eat a lot of resources, especially on disk access. It
installs itself as a service under launchctl, and this prevents it from being
killed directly. However the process can be suspended. It's like encasing an
ancient, unkillable evil in ice and hope that no one ever finds it and thaws
that ice.

It would be nice to have something that re-enables it automatically during a
time of day in which I am not expected to be productive, because Security wants
it to run. One day we should investigate the brainwashing powers of the Sophos
virus since it can influence the minds of those in a security department. Such
power!