optimize helm
Helm is very slow on my machine for large-ish files. This follows a bugfix which
should fix the slow-downs I see quite a bit. It seems to halve the amount of
time I spend waiting on it. But a 4 second delay down to 2 is still pretty
nasty. I have attempted to use helm in a frame, and while this completely fixes
the performance issues, it isn't ready since it doesn't display
properly (sometimes not at all) on multiple displays. My attempts at using a
frame are in this commit, but not attempts on fixing the frame positioning for
multiple displays.