consolidate zsh settings into zshrc
Due to the order of inclusion, using zshenv to setup the PATH is problematic,
because I cannot put things at the start of the path. zshenv goes first, which
means it loses when other things want to go at the beginning of the PATH. Moving
everything in zshenv to zshrc (which goes last or last-enough) means I can put
things on any end of the PATH or even in the middle.