use option/alt as the meta key
The option/alt key on MacOS is useful for a variety of reasons, and Emacs
usually treats it as meta. On MacOS I can use it to type characters with
diacritical marks. However alt has no bindings on Emacs unless I set them. To
this end I am disabling command as meta and leaving it be alt. Now I have to
form new muscle memory, but that's okay. It also means I can use the command key
as hyper, which reflects the operating system-like semantics of the MacOS
bindings. Perhaps I can find a good super key now.