pipe buffer into flow type-at-pos to keep current edited version
The prior version would indicate that type-at-pos should read from the file
directly, which would often be out of date with the buffer. Piping in the buffer
directly took some work, as documentation for type-at-pos for pipes is
incomplete. This could probably be refactored to use the shell-command-to-string
function. I was trying to use shell-command-on-region which took the current
buffer and piped it into stdin, but this seemed to have no effect. Now it's an
escaped string sent via echo. It's not optimal but the speed seems fast enough
for my purposes and most importantly, it works. We can optimize later.