add code-injector support
Safari 13 disables custom extensions. Boo! I looked at Firefox and found a
solution. I can massage the existing data structure a little bit to make it fit.

Currently I have to use a custom solution to load all of my data. The import
doesn't seem to work, and I want to be able to rapidly iterate on these when
necessary. This means rerunning it in Firefox on each restart, but that
shouldn't happen often, and the web-ext CLI app holds promise in mitigating that
as well. Not bad for a first pass though!