fix issue where navigation causes dark theme to be lost
The onLoad flag causes the code injection to wait until the page is done
loading, so it makes for an obnoxious white flash as content comes in. These
changes provide some pre-styling to prevent that. The tradeoff is that all pages
will be dark themed now for better or worse. I might need to find another way to
apply this going forward. Perhaps I can aggregate all of the matchers and load
them into global.json so it only themes the sites I already support.