clean up javascript files and prepare for working flycheck checkers
There are problems with flycheck checking at the moment. lsp-ui disables
automatic checking, and after testing I have found this is probably for a good
reason (it gets stuck after saving). I don't have a solution for now, but I
should be able to apply it without re-wiring anything now.