add helpers to scrape html info from sites as org-mode
With crawl/dwim-from-url I can paste a URL and the system will download it and
convert the HTML into .org. It then inserts this as a subheading, with the
incoming org headings demoted to fit under this new heading.

I don't think pandoc supports images. Even if it did, I would still be relying
on the original source to provide those images. I would like to avoid that.