follow experimental code-injector changes
I hopped on the develop branch and did a number of fixes to it. From
that I needed to update to the new "v3" of the data format. I have the
old format in place as well, since I don't know if the author will leave
v3 where it is.

With the new branch I no longer need a build step, so inject-browser.sh
is now obsolete.