make nix install and upgrade work
Nix looks for presence of itself on a new system, but that fails with my
dotfiles in place. It will detect itself in the shell rc files and bail
out. This is the wrong thing to do since I have set it appropriately.
Instead we break the string up to thwart Nix's detection of itself.

Changes are also made to make the install work on macOS Catalina.

The upgrade should now work all of the time. Before it was an untested
snippet.