switch to home-manager, install openconnect-sso
This gets openconnect-sso working in a really messy way. Though it
works! There's still a lot of cleanup to do, and perhaps some bug fixing
as well. Other .nix files have been committed because they are just too
outstanding at this point.

This also puts us on home-manager, which is a much cleaner way of
managing the home setup, and is now part of the official nix
documentation.

These changes allow me to continue to use VPN at my workplace.