fix DNS recovering when leaving the VPN
This seems to be related to macOS 11.3 changes I am not aware of.
Calling the dnsReset function twice seems to do the trick, but is only
needed when running vpn-nwea directly. The nwea-clock-in script doesn't
have this issue. Perhaps it is a timing issue? I haven't been able to
figure it out yet, so the hacks grow.