drop repeat keys in evil-snipe
f/F/t/T would be overridden to repeat a search, which has been
disruptive to my usage of those keys (I expect them to be one-shot, not
a new mode). It took a while to reach this configuration since I didn't
understand the documentation on first read.

This also brings about use-package for evil-snipe and the other settings
I have with it, which might help with load times and such.