stop using emoji during font resizing
This is tested and works great. However, an issue remains at large
regarding the font bold/italics settings when doing a font resize.
Fortunately, when using doom--big-font-mode, the original font is
restored. I haven't checked to see if bugs exist for this, or if an
update will clear this bug out.