install macports for aarch64, install font properly
The Source Code Pro font exists as a Nix derivation already, but only
works for Linux setups. I created an overlay based on that derivation
which places the font in the proper directory for macOS. I need to
confirm that my symlinking done in neon.nix was also needed, or if that
symlinking alone would've taken care of the problem. Once done I can see
about submitting a pull request for the font if it is needed.

I also add Emacs support with a properly symlinked ~/Applications
directory. It crashes upon startup in various ways. I don't have
debugging information so this might need to be moved back to a homebrew
setup. My Emacs configuration is non-trivial, so this brittle step-back
is highly undesirable.