enable dark mode in Facebook Messenger
This enables dark mode via JavaScript - there is a "light mode" class on
the html element, and quick experiments showed that changing the "light"
to "dark" worked for flipping this mode. It's probably experimental,
because their UI still doesn't have a setting for it. That said, the
mode works well except for individual message backgrounds, which we fix
here.