fix line number shifting
The moment I ran this, it fixed a shifting issue where the current line
would shift slightly to the right. Scrolling around would make a silly
wave effect.

This is due to a bold being set in the current line number. Some other
effect should be used, because bolding the font means that the spacing
is shifted slightly. Bolding also just generally doesn't work well in
dark themes, and that's where I live.