update flakes and add fuse support
This does require disabling the SIP on macOS - a generally undesirable
state.  Still, it could be done temporarily, and so having the host
state ready for it will save cycles if necessary.

This does require some patching to `e2fsprogs` and they don't feel safe.
I should learn more and/or contribute these.