bump nextcloud, declare Nextcloud LDAP config
This bumps nextcloud in the hopes that it will be more resilient when a massive
phone sync is happening.

The LDAP configuration got out of sorts during the upgrade, so we've made it
work declaratively as it should've been early on.  Now we have notes for the
tools to detect and set these things in the future too.